#926: Split played vs unplayed layout engines into different files
