+++ CourageScores/ClientApp/src/components/tournaments/layout/shared.ts
@@ -1,0 +1,13 @@
+export interface IMnemonicAccumulator {
+    next(): string;
+}
+
+export function getPrefixIncrementingMnemonicCalculator(prefix: string): IMnemonicAccumulator {
+    let index: number = 0;
+    return {
+        next(): string {
+            return prefix + (++index);
+        }
+    };
+}
+
